Move language selector to header
To get more in line with especially Studio redesign,
this moves the language selector from the settings page
to the header and styles it accordingly
