+++ src/main/MainContent.tsx
@@ -23,7 +23,6 @@ import { selectHasChanges as selectSubtitleHasChanges } from "../redux/subtitleS
 import { useTheme } from "../themes";
 import Thumbnail from "./Thumbnail";
 import Cutting from "./Cutting";
-import LanguageSwitcher from "./LanguageSwitcher";
 
 /**
  * A container for the main functionality
@@ -131,7 +130,6 @@ const MainContent: React.FC = () => {
     } else if (mainMenuState === MainMenuStateNames.keyboardControls) {
       return (
         <div css={[mainContentStyle, keyboardControlsStyle]} role="main">
-          <LanguageSwitcher />
           <KeyboardControls />
         </div>
       )
